apps/whale-monitor/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
websocket-client==1.7.0
